Separte the TCs and mark other one as excluded
