+++ config/config.yml
@@ -62,3 +62,4 @@ excluded_tests:
     - tests/functional/glusterd/test_shared_storage.py # Reboot hangs redant flow in this TC.
     - tests/functional/afr/test_brick_process_not_started_on_read_only_node_disks.py # Inactive TC.
     - tests/functional/snapshot/test_snapshot_create.py # Fails in upstream. Succeeds in downstream env
+    - tests/functional/afr/test_client_side_quorum_fixed_with_cross2.py # TC is bogus, as setting quorum count: 2 for replica 2 volume and bringing one brick down will give transport errors
